api_gateway_lambda_dynamo/requirements.txt
aws-cdk.core
aws-cdk.cloudformation_include
aws_cdk.aws_lambda
aws_cdk.aws_logs
aws_cdk.aws_dynamodb
boto3
aws_cdk_.aws_apigateway